services/user/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
python-jose==3.3.0
passlib==1.7.4
asyncpg==0.28.0
psycopg2-binary==2.9.7
redis==5.0.0
faststream==0.2.5
faststream[kafka]==0.2.5
fast-depends==2.2.0
aiohttp==3.8.5
python-multipart==0.0.6
bcrypt==4.0.1
email-validator==2.0.0
python-dotenv==1.0.0
kafka-python==2.0.2
PyJWT==2.8.0